controller/supervisor: move supervisor to separate package
